+++ src/App.jsx
@@ -1,6 +1,5 @@
 import { Outlet, Route, Routes } from 'react-router-dom'
 import { Home } from './components/views/Home'
-import { ThemeToggle } from './components/ThemeToggle'
 import { Navbar } from './components/Navbar'
 
 function App() {
