+++ src/components/Navbar.jsx
@@ -1,4 +1,5 @@
-import { ThemeToggle } from "./ThemeToggle";
+import { ThemeToggle } from "./ThemeToggle/ThemeToggle";
+
 
 
 export const Navbar = () => {
