--- src/components/ThemeToggle.jsx
@@ -1,27 +1,0 @@
-import { useState, useEffect } from "react";
-
-export const ThemeToggle = () => {
-  const [theme, setTheme] = useState(localStorage.getItem('theme') || 'light');
-
-  useEffect(() => {
-    if (theme === 'dark') {
-      document.documentElement.classList.add('dark');
-    } else {
-      document.documentElement.classList.remove('dark');
-    }
-    localStorage.setItem('theme', theme);
-  }, [theme]);
-
-  const toggleTheme = () => {
-    setTheme(theme === 'dark' ? 'light' : 'dark');
-  };
-
-  return (
-    <button
-      onClick={toggleTheme}
-      className="p-1 w-6 h-6 bg-gray-200 dark:bg-gray-800 text-m rounded-full flex items-center justify-center content-center"
-    >
-      {theme === 'dark' ? '🌝' : '🌚'}
-    </button>
-  );
-};
